Doing some formatting and moving some stuff
